pyproject.toml
[tool.poetry]
name = "etl-dengue"
version = "0.1.0"
description = ""
authors = ["Rhanyele Marinho <rhany.marinho@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pandas = "^2.2.2"
sqlalchemy = "^2.0.29"
logging = "^0.4.9.6"
psycopg2 = "^2.9.9"
streamlit = "^1.33.0"
plotly = "^5.21.0"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
